+++ SEPADirectDebit/src/main/java/com/braintreepayments/api/SEPADirectDebitRequest.java
@@ -14,6 +14,7 @@ public class SEPADirectDebitRequest {
     private SEPADirectDebitMandateType mandateType = SEPADirectDebitMandateType.ONE_OFF;
     private PostalAddress billingAddress;
     private String merchantAccountId;
+    private String locale;
 
     /**
      * @return The account holder name
@@ -109,5 +110,24 @@ public class SEPADirectDebitRequest {
      */
     public void setMerchantAccountId(@Nullable String merchantAccountId) {
         this.merchantAccountId = merchantAccountId;
+    }
+
+    /**
+     * @return A locale code to use for creating a mandate.
+     */
+    @Nullable
+    public String getLocale() {
+        return locale;
+    }
+
+    /**
+     * Optional.
+     * @param locale A locale code to use for creating a mandate.
+     *
+     * @see <a href="https://developer.paypal.com/reference/locale-codes/">Documentation</a>
+     * for possible values. Locale code should be supplied as a BCP-47 formatted locale code.
+     */
+    public void setLocale(@Nullable String locale) {
+        this.locale = locale;
     }
 }